requirements.txt
anyio==4.4.0
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
et-xmlfile==1.1.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
pandas==2.2.2
idna==3.7
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-telegram-bot==21.4
pytz==2024.1
requests==2.32.3
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.2
